cran_indexer/requirements.txt
PyYAML==3.11
elasticsearch==1.6.0
urllib3==1.10.4
wsgiref==0.1.2
